Add query tests
These are run by 'tree-sitter test'. The 'highlights.scm' won't work
out of the box for any editor but they serve as a starting point.

Testing with queries allows you to do essentially the same thing as the
unit tests under 'test/corpus/'. Query tests also allow you to assert
that the grammar is structured in a way that allows consuming
applications to write queries the ways you want. For an example of
something that this catches, see the parent commit.