Allow for alternate comment characters.
Git allows the [comment
character](https://git-scm.com/docs/git-config#Documentation/git-config.txt-corecommentChar)
for commands such as commit and tag to be changed. The original
tree-sitter grammar only expects `#` to be used and fails to highlight
commit messages if an alternate character is used. This change adds a
fixed list of additional characters, copied from [neovim].

Resolves #16.

[neovim] https://github.com/neovim/neovim/blob/2b21c9c23f889555dd93ba508a0f787b25ed9bb4/runtime/ftplugin/gitcommit.vim#L21